feat: continuously insert data

+++ src/main/resources/postgresql_schema.sql
@@ -13,11 +13,12 @@ CREATE TABLE singers (
 
 -- When converting to Spanner:
 --   * singer_id has int4 type, which must be converted to int8 in Spanner
+--   * SERIAL type does not exist in Spanner
 --   * album_title UNIQUE must be converted to an UNIQUE index in Spanner
 --   * FOREIGN KEY relation can be converted to Spanner INTERLEAVED table
 CREATE TABLE albums (
   singer_id     int,
-  album_id      bigint,
+  album_id      SERIAL,
   album_title   VARCHAR UNIQUE,
   PRIMARY KEY (singer_id, album_id),
   CONSTRAINT fk_singers
@@ -26,12 +27,13 @@ CREATE TABLE albums (
 );
 
 -- When converting to Spanner:
+--   * SERIAL type does not exist in Spanner
 --   * song_data json type must be converted to jsonb in Spanner
 --   * FOREIGN KEY relation can be converted to Spanner INTERLEAVED table
 CREATE TABLE songs (
   singer_id     int,
-  album_id      bigint,
-  song_id       bigint,
+  album_id      int,
+  song_id       SERIAL,
   song_name     VARCHAR,
   song_data     json,
   PRIMARY KEY (singer_id, album_id, song_id),
